+++ authcode/src/main/java/com/stargatex/auth/authcode/configs/auth/AuthConfiguration.kt
@@ -3,7 +3,7 @@ package com.stargatex.auth.authcode.configs.auth
 import com.stargatex.auth.authcode.configs.client.ClientAuthConfig
 import com.stargatex.auth.authcode.configs.client.ClientSecretConfig
 import com.stargatex.auth.authcode.model.auth.AuthRequestOptionalConfig
-import com.stargatex.auth.authcode.model.ocid.OidcConfig
+import com.stargatex.auth.authcode.model.oidc.OidcConfig
 
 /**
  * @author Lahiru J (lahirujay)
