+++ authcode/src/main/res/raw/oidc.json
